feat(ui): display XP with decimal precision
- Convert XP from integer to float in backend data processing
- Update profile and stats templates to show XP with 1 decimal place
- Improve CSS formatting in profile template

+++ app.py
@@ -1998,7 +1998,7 @@ def stats_page():
         for row in agents_response.data:
             name = row['name']
             faction = row.get('faction', 'Wanderer')
-            xp = int(row.get('xp', 0))
+            xp = float(row.get('xp', 0))
             
             registry[name.lower().strip()] = {
                 'name': name, 'faction': faction, 'xp': xp
